requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.27
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
importlib_metadata==8.5.0
inflection==0.5.1
Markdown==3.7
packaging==25.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.13.2
tzdata==2025.3
uritemplate==4.1.1
zipp==3.20.2
